requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
build==0.5.1
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
cryptography==38.0.3
html5lib==1.1
httplib2==0.21.0
idna==2.10
Naked==0.1.32
oauthlib==3.2.2
packaging==20.9
pep517==0.10.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==22.1.0
pyparsing==2.4.7
python-dotenv==0.21.0
pywatchman==1.4.1
PyYAML==6.0.1
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
shellescape==3.8.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.6
webencodings==0.5.1